chore(testworkflows): extract code for executing parallel operations
